chore: add changelog (#410)
So far, I've only been tracking user-impacting changes. This was to
reduce overhead and make it easy to share with users.

Also, the format is intentionally spartan. If updating this file becomes
a burden, we just won't do it. So, I don't think it's a good idea to do
any fancy stuff, at least for now.

The only above-minimal-effort thing that may be worth it is to identify
breaking changes so that we can communicate it to users.

Signed-off-by: Akhil Velagapudi <4@4khil.com>