chore(nox-ecs): simplify client world init
A once cell was being used to lazily load buffers to client, but this
was unnecessary because there was already a mechanism for doing this:
dirty component tracking. If we track all components as dirty during
shared world init, then all buffers will be copied from host to client.

Also, removed loaded components tracking as it doesn't really serve a
purpose. We load all components from client to host every time.

Finally, moved host -> client and client -> host buffer transfers to
WorldExec to be consistent. Previously, host -> client was done in Exec
and client -> host was done in WorldExec.

Signed-off-by: Akhil Velagapudi <4@4khil.com>