feat(nox-ecs): make pushing to history faster
Do the least amount of work necessary to store history: cloning the
buffers. This requires some refactoring as we need to separate metadata
(which is the same every tick) from the buffers themselves.

Also, the polars DataFrame is only built when requested because it's too
slow to do inline. Eventually, we can move this to an asynchronous
process.

Signed-off-by: Akhil Velagapudi <4@4khil.com>