feat(nox-py): remove elodin-types dependency
This dependency is pretty annoying as you have to pull in all of
protobufs/. The only reason it was needed is to provide unique names to
each sample in a monte carlo run. The name of the sample directory is
guaranteed to be unique (and actually is the sample_number currently),
so it's sufficient as a unique pytest id.

Signed-off-by: Akhil Velagapudi <4@4khil.com>