Improve sun sensor, magnetometer, and reaction wheel simulations in cube-sat example (#516)
This PR adds more realistic sensor models to the cube-sat example. Previously, we used dummy sensors with noise. I also added a basic reaction wheel model that provides a limit on angular momentum.

The reaction wheel model is based on the NASA 42 model, with oscillations and friction removed. The course-sun-sensor model is from Basilisk, and the magnetometer model is from both NASA 42 and Basilisk