feat(drone): add attitude estimation, rate + angle controllers (#482)
* feat(drone): add gyro with 40 Hz lp filter

The gyro sample rate is 900 Hz, which is close to the 1 kHz rate used by
Ardupilot.

Ardupilot's filter strategy is actually a bit more complicated. The raw
gyro data is read at 8 kHz but averaged down to 1 kHz in a fast loop.
The accelerometer data is read at 1-4 kHz, goes through a 1st order
low-pass filter at 188 Hz, and then gets averaged down to 1 kHz. This is
all before going through the "main" set of low-pass and notch filters at
1 kHz. IDK why it's so complicated ...

I'm not going to add additional filters until it's necessary. The
low-pass one seems to be the most important because it's enabled by
default. All the other ones are disabled in the "pre-tuned"
configuration. You're supposed to conduct a bunch of tuning flights in
order to set up the notch filters as those are somewhat specific to the
motor/prop/frame configuration.

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* fix(drone): align motor positions to bevy x,z axes

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* fix(drone): gyro ang vel from world -> body

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* feat(drone): add rate controller

We can now define setpoints for roll, pitch, and yaw rates. The PID
response looks too good for pitch and roll. Some contributing factors
are:
- Less inertia along those axes
- No delay in motors achieving desired thrust

This feels pretty unrealistic; we really need to characterize the motor
delay by measuring the settle time.

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* feat(drone): move params to separate file

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* feat(drone): add 1st order lpf

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* feat(drone): move sensors to separate file

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* feat(drone): add accelerometer with 10 Hz lpf

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* feat(drone): add drag, pure grav accel component

The "grav_accel" component is only used for debugging purposes as it
represents the ideal case for the accelerometer data (when there is no
external acceleration).

In an equilibrium state, the "raw" acceleration should match the "pure"
gravitational acceleration.

Also, add a simple aerodynamic drag system so that constant motor thrust
eventually leads to net 0 external acceleration.

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* feat(drone): add IMU noise

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* feat(drone): add first-order motor thrust response

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* chore(drone): split into sim.py, main.py

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* feat(drone): add magnetometer

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* feat(drone): add mekf attitude estimation

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* fix(drone): Y-up -> Z-up

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* feat(drone): adjust accel covariance dynamically

When there is a significant amount of lateral acceleration, the
accelerometer becomes less useful at determining attitude. So, increase
the covariance dynamically based on a "health" factor which approaches 0
during dynamic flight conditions.

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* fix(drone): convert magnetometer from NED to ENU

east, north, up (ENU) -> north, east, down (NED)

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* feat(drone): move rate controller to control.py

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* feat(drone): add angle controller

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* feat(drone): make motor mixing more accurate

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* feat(drone): add T,E,D filters for rate controller

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* feat(drone): remove unnecessary "desired_thrust"

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* feat(drone): specify all PID gains in params file

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* fix(drone): address PR comments

Signed-off-by: Akhil Velagapudi <4@4khil.com>

---------

Signed-off-by: Akhil Velagapudi <4@4khil.com>