Add a skeleton for the STM32 fw (#709)
* feat(fsw): add a skeleton for the STM32 fw

Currently, all it does is generate a PWM signal using DMA to vary the
duty cycle. It should be pretty easy to add dshot on top of this.

The use of stm32-hal2 may seem a bit odd as it has some clear downsides:
- Not the most popular HAL (the stm-rs one has more downloads)
- No first-party runtime integration (unlike embassy)
- Very few high-level APIs

But I still think it's the right choice because it doesn't try to hide
the underlying complexity from the user. The APIs are verbose, but tend
to mirror the the reference/user manuals pretty well. I literally tried
all the other HALs, and this one was the easiest to debug (by far).

The biggest advantage IMO is that it's well maintained. It has regular
releases to crates.io (unlike embassy-stm32 OR stm32h7xx-hal). It's
already on embedded-hal 1.0 (unlike stm32h7xx-hal). The maintainer is
very responsive to issues + PRs. I'm pretty confident that we can
upstream improvements if needed.

I'm excluding this package from the workspace because of 2 reasons:
1. Workspaces don't work well for packages that only build for specific
   targets.
2. Maintaining a separate lockfile allows us to audit dependencies
   better.

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* feat(fsw): use a mutable buf for DMA

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* feat(fsw): add helpers to avoid redundant args

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* feat(fsw): add safe DMA API

Signed-off-by: Akhil Velagapudi <4@4khil.com>

---------

Signed-off-by: Akhil Velagapudi <4@4khil.com>