feat(db): accept msgs over UDP (#908)
This makes it trivial to mirror entire db instances over a UDP unicast.
TODO: add support for mirroring metadata. Currently, the "replica" needs
to be seeded with metadata info independently. But this isn't much of an
issue because you can just use the same config for starting the replica.

I'll add more convencience stuff in a follow-up PR.

Signed-off-by: Akhil Velagapudi <4@4khil.com>