fix(os): enable hardware.graphics to symlink libcuda.so to /run/opengl-driver/lib/
This is the "standard" way for nixos packages to discover libcuda. Rpath
patching to an exact cuda lib doesn't work because cuda is tightly
coupled with the kernel module. So, they're all patched to look for
shared libraries in "/run/opengl-driver/lib/" and
hardware.graphics enables an activation script to place the necessary
symlinks in "/run/opengl-driver/lib/". With this change, setting
LD_LIBRARY_PATH is no longer required for native nix packages. It's
still needed for python scripts.

Signed-off-by: akhilles <akhilvelagapudi@gmail.com>