Add pycuda, fix libcuda.so not found issue (#42)
* fix(aleph-os): add user to video group to use cuda without sudo

This also fixes the issue where libcuda.so couldn't be found because
LD_LIBRARY_PATH is in fact properly set for users and for root. The
problem is with sudo clearing this environment variable.

Signed-off-by: akhilles <akhilvelagapudi@gmail.com>

* fix(os): enable hardware.graphics to symlink libcuda.so to /run/opengl-driver/lib/

This is the "standard" way for nixos packages to discover libcuda. Rpath
patching to an exact cuda lib doesn't work because cuda is tightly
coupled with the kernel module. So, they're all patched to look for
shared libraries in "/run/opengl-driver/lib/" and
hardware.graphics enables an activation script to place the necessary
symlinks in "/run/opengl-driver/lib/". With this change, setting
LD_LIBRARY_PATH is no longer required for native nix packages. It's
still needed for python scripts.

Signed-off-by: akhilles <akhilvelagapudi@gmail.com>

* feat(aleph-os): add pycuda support

Add some checks to cudainfo that exercise pycuda.

Signed-off-by: akhilles <akhilvelagapudi@gmail.com>

---------

Signed-off-by: akhilles <akhilvelagapudi@gmail.com>