build(nix) - enable fallback for nix build
nix - If binary caches are are offline, then nix builds will fail unless
the flag --fallback is enabled. This also breaks the deploy script if
any binary cache is offline.

This commit enables the fallback feature by default, so that nix will
"fall back" to building from source if a binary cache is missing.