feat(aleph) - allow custom save location for db (#131)
aleph - in elodin-db module, the database will now be saved as a
sub-folder of /db. The subfolder name is customizable and can be passed
as an argument to the service, where:

systemctl start elodin-db@NAME.service

will result in a database saved to /db/name

The parent folder (/db) is also customizable as a variable in the
elodin-db nix module.

Co-authored-by: labrnth <labrnth@labooth.net>