feat(docs): Add a means of testing public docs.
The public docs are written in a semi-literate programming style.
Semi-literate in the sense they aren't written in a Knuth Literate
Programming style, which is a complex and comprehensive system for
creating multiple program files from multiple document files. The
semi-literate programming uses a simpler approach, one markdown file to
create one code file.

Semi-literate programs differ from doc-tests because code blocks in
doc-tests are isolated: one code block has no bearing on another.
Instead a semi-literate program blocks are expected to accrue from top
to bottom. In a way, it's like writing a program but flipping it
from default code and exceptional comments to default comment and
exceptional code.

This commit adds a literate-tests crate internal to docs that can be
directed to extract Rust or Python code from markdown files and then
test, exercise, or check formatting of those of those files. In this
way, the aim is to bring the public docs under our regular CI test
regime, so that they will be flagged when changes break them.