Implement wait_for_write functionality. (#277)
* perf: Use one string.

Don't allocate a string for each component.

* hack: The going got good. Stop.

* test: Hash msg.

* hack: Check vtable registry.

It was empty.

* hack: Never wait.

Like before.

* Revert "hack: Check vtable registry."

This reverts commit 6a8a37225bc29dbd7cf17b7407f6f5e403df4074.

* feat: Try wait for write.

* fix: Avoid subtract overflow.

When I ran the rocket example once, I got this:
```
thread 'main' panicked at libs/elodin-editor/src/ui/plot/widget.rs:1204:13:
attempt to subtract with overflow
```
It was not reproducible.

* fix: Plot numerical overflows.

Tend to present themselves when the sim is not writing immediately, so
lots of things did not work on the first go around.

* refactor: Make run_time_step Option<Timestep>; no waits on None.

Do not wait unless run_time_step has been set. It's default value is now
None.

* feat: Add `Component::from_pair(entity_name, component_name)`.

Avoid a string alloc for this hash computation.

* feat: Add "wait_for_write" support.

Some general clean up. Tried to deal with computables outside of hot
loop. Keep IDs of what we need up front.

* feat: Add rocket-client.cpp example to libs/db.

* style: Reformat.

* feat: "wait_for_write" should work with headless run.

Not tested, but had to remove that `loop {}` hack.

* style: Satisfy fmt and clippy.

* test: Fix failing doctest.

* doc: Update python dev to run from root.

* style: Reformat.

* hack: Remove dbg!

* excise: Remove wait_for_write on headless.

* doc: Update `run_time_step` for new default value.