refactor: Check for signals in Rust not Python.
We run our JAX World in a thread not limited by the Python Global
Interpreter Lock (GIL). However, prior to this commit, we acquire the
GIL each time we check if we should cancel by checking for Unix signals
using `py.check_signals()`. This commit changes our signal checking to
use a Rust library and sets an atomic bool flag that we then check for,
so we do not not need to acquire the GIL unless we're actually running a
Python function.