Support SITL Python module (#323)
* ai: Add is_canceled function to World.run.

* chore: Add other is_canceled necessaries.

* refactor: Avoid Arc if possible.

* feat: Add is_canceled to Exec.run().

* feat: Add db_path to world.run().

* fix: Box the DB so it doesn't spread alignment issue.

* feat: Use db_path if provided.

* refactor: Better error report for time out of bounds.

* feat: Add args to elodin-db.

Accept the remaining arguments:

$ lua script.lua 1 2 3

* feat: Add db_path and interactive to World.run().

* refactor: Check for signals in Rust not Python.

We run our JAX World in a thread not limited by the Python Global
Interpreter Lock (GIL). However, prior to this commit, we acquire the
GIL each time we check if we should cancel by checking for Unix signals
using `py.check_signals()`. This commit changes our signal checking to
use a Rust library and sets an atomic bool flag that we then check for,
so we do not not need to acquire the GIL unless we're actually running a
Python function.

* chore: Make it compile.

* fix: Setup Lua arg unconditionally.

* test: Refactor fix.

* style: Reformat.

* style: Reformat.

* feat: Add `post_step` argument for SITL integration.

* style: Reformat.