feat(aleph) - allow unique db save location
Commit 345f9b9b6359591a41348ed890a9d90945e2abde created a service
where elodin-db would be saved as a sub-folder of /db. By default,
elodin-db starts on boot and saves to a default db named.. well
'default'.

This commit adds an option to save in a unique DB upon each boot. This
is useful if you are using an alternate clock source, such as
CLOCK_MONOTONIC. This setting also configurable with the following nix
option: `services.elodin-db.dbUniqueOnBoot`. It defaults to true.